style(lytfs): rename env var to LYTFS_GITHUB_TOKEN
Rename the env var used by lytfs to LYTFS_GITHUB_TOKEN to adhere to the
convention that env vars are generally prefixed by the program name that
uses them.

Formerly this was yagfuncdata, but after review this changed to lytfs.
Furthermore this token is exclusively for GitHub, thus it should also be
named GITHUB_TOKEN instead of AUTH_TOKEN.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>