build: add a "simple" release script
Add a "simple" release convenience script that takes out a lot of
headaches revolving around releasing a new Go Module version.

It assumes that gh-cli is installed to also create a new release, but as
we do that as our very last step, we can safely exit out at that point
and redirect to the web interface, should gh-cli not be installed.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>